Change README.md, fix Session model, remove unused resources, change ConfirmationActivity.kt, fix proguard-rules.pro, change gradle dependencies, optimize layouts

+++ keyrisdk/src/main/java/com/keyrico/keyrisdk/entity/AccountData.kt
@@ -5,7 +5,7 @@ import com.google.gson.annotations.SerializedName
 data class AccountData(
 
     @SerializedName("ttl")
-    val ttl: Long,
+    val ttl: Double,
 
     @SerializedName("id")
     val id: String
